changes according to pauls comments
